Moved pre checks to the top of the expression tree instead of the bottom
This helps in a few ways: first, it allows common subexpression
elimination to work across pre boundaries, reduces the number of pre
expressions, and on the Plexil side it removes the problem of not having
a good way to refer to the current values of variables. Now a translator
can make that decision itself.