Static typing in IL, UNKNOWN now works for numeric values in Lustre
This change was huge and painful, but the upshot is twofold:

- The IL is much more strict about types: if it isn't provided
statically, even for lookups, it is supposed to yell at you
- With this newfound type strictness, the Lustre translation can now
watch out for numeric values. That allows for unknown numerics.

Since we can't enumerate all numbers, they have to be a pair of the
number and whether it's known. Unfortunately, Lustre doesn't give us
data structures to pack these into, so we have to handle splitting and
combining them at translation time.

(Tuples exist, but not as first-class citizens: they cannot be passed as
arguments to nodes, they are *only* used for multi-assignments and
multi-returns, basically. At translation, we return tuples, but to
operate on them the tuples must be split and re-constituted. So (1,true)
+ (2,true) has to become (1+2,true and true) because something like
p_add((1,true), (2,true)) is not allowed. We are allowed to use tuples
to make variable assignments look sensible, so the "value" part and the
"known" part of a numeric variable are both assigned with the same
equation.)