Removed all asserts from Lustre translation
The problem was that JKind slices out asserts when they don't apply to
the property being checked. That's okay, since the property is violated
without violating the assert, but it means that we don't get the values
that make the assert valid. We need those to extend the test case,
because otherwise we have to either guess, which means we could be
creating a test that violates asserts, or we need to get a solver to
re-solve it for us. 

All of that is dumb. So instead, we're just doing the simple thing: all
inputs are now completely unconstrained. When a constraint is needed, we
make a local variable with the constraints and everything that needs to
read that input instead reads the constrained version.