Test case generation can now be done in one step
The option `--incremental-search` now starts up incremental test case
generation for each PLX file passed in for translation. The resulting
Lustre test cases are translated to PLEXILScript format in-memory. With
`--lang plexil`, those scripts will be output to the output directory.

In addition, the `--compliance` option can be used to compliance check
the scripts versus `plexiltest`. 

To avoid pretty-printing the PLX file too when we do this, I changed the
pretty-printing functionality to be under `--lang pretty`.