Updated the possible command handle values and implemented aborts
At some point in the last couple years, the list of possible command
handles changed. Along with it, the semantics for command aborts also
changed (previously, an abort was acknowledged with a specific handle
response, now it is implemented as an internal "ack" variable, similar
to Update nodes.) Since we had never implemented aborts in the first
place, this seemed like the time to do it. 

We were having trouble because JKind was generating tests with the old
handles, and official PLEXIL was rejecting them as invalid. Now that we
are all synced up, JKind can no longer generate illegal values.